+++ Sources/SakeKit/String+Extras.swift
@@ -3,7 +3,7 @@ import Foundation
 extension String {
     
     static func randomString(length: Int = 10) -> String {
-        let letters : NSString = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+        let letters: NSString = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
         let len = UInt32(letters.length)
         var randomString = ""
         for _ in 0 ..< length {