+++ Sources/SakefileDescription/Sake.swift
@@ -19,9 +19,13 @@ public final class Sake {
 public extension Sake {
     
     public func run() {
-        tasksInitializer(tasks)
         var arguments = CommandLine.arguments
         arguments.remove(at: 0)
+        run(arguments: arguments)
+    }
+    
+    func run(arguments: [String]) {
+        tasksInitializer(tasks)
         guard let argument = arguments.first else {
             print("> Error: Missing argument")
             exit(1)
@@ -34,7 +38,7 @@ public extension Sake {
                 exit(1)
             }
             do {
-                try run(task: arguments[1])
+                try runTaskAndDependencies(task: arguments[1])
             } catch {
                 print("> Error: \(error)")
                 exit(1)
@@ -53,13 +57,24 @@ public extension Sake {
             .joined(separator: "\n"))
     }
     
-    fileprivate func run(task taskName: String) throws {
+    fileprivate func runTaskAndDependencies(task taskName: String) throws {
         guard let task = tasks.tasks.first(where: {$0.name == taskName}) else {
             return
         }
-        try task.dependencies.forEach({ try run(task: $0) })
+        tasks.beforeAll.forEach({ $0(utils) })
+        defer { tasks.afterAll.forEach({ $0(utils) }) }
+        try task.dependencies.forEach { try runTask(task: $0) }
+        try runTask(task: taskName)
+    }
+    
+    fileprivate func runTask(task: String) throws {
+        guard let task = tasks.tasks.first(where: {$0.name == task}) else {
+            return
+        }
         print("> Running \"\(task.name)\"")
+        tasks.beforeEach.forEach({$0(utils)})
         try task.action(self.utils)
+        tasks.afterEach.forEach({$0(utils)})
     }
     
 }