+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## 0.1.0 - (next version)
 
 ### Added
+- Hooks https://github.com/xcodeswift/sake/pull/9 by @pepibumur
 - Xcode project generation https://github.com/xcodeswift/sake/pull/4 by @pepibumur.
 - Homebrew support https://github.com/xcodeswift/sake/pull/5 by @pepibumur.